Class "Population"

Contains an instance of a population including a description of their detectability in the form of an object of class Detectability.

Slots

region.name

Object of class "character"; the name of the region object.

strata.names

Object of class "character"; the names of the strata.

N

Object of class "numeric"; the number of individuals/clusters.

D

Object of class "numeric"; the density of individuals/clusters.

population

Object of class "data.frame"; the locations of individuals/clusters and any population covariates.

detectability

Object of class "Detectability"; describes how easily the individuals/clusters can be detected.

Methods

plot

signature=(object = "Line.Transect"): plots the locations of the individuals/clusters.

See also