Class "Simulation.Summary"

Class "Simulation.Summary" is an S4 class containing a summary of the simulation results. This is returned when summary(Simulation) is called. If it is not assigned to a variable the object will be displayed via the show method.

Methods

show

signature=(object = "Simulation.Summary"): prints the contents of the object in a user friendly format.