Update cookies preferences

Class "Survey.LT" extends class "Survey"

Class "Survey.LT" is an S4 class containing a population and a set of transects.

Slots

transect

Object of class "Line.Transect"; the line transects.

perpendicular.truncation

Object of class "numeric"; the maximum distance from the transect at which animals may be detected.

See also