<<<<<<< HEAD

<<<<<<< HEAD
S4 generic method to add a hotspot to the density grid
=======
S4 generic method to add a hotspot to the density grid
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

Uses a Gaussian decay around a central location to add a hotspot to the density grid.

Usage

add.hotspot(object, centre, sigma, amplitude)

<<<<<<< HEAD
# S4 method for Density
=======
# S4 method for class 'Density'
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1
add.hotspot(object, centre, sigma, amplitude)

Arguments
<<<<<<< HEAD

object

a Density-class object

centre

an x,y-coordinate giving the centre of the hotspot

sigma

a value giving the scale parameter for a gaussian decay

amplitude
=======

object

a Density-class object

centre

an x,y-coordinate giving the centre of the hotspot

sigma

a value giving the scale parameter for a gaussian decay

amplitude
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

the height of the hotspot at its centre

Value
<<<<<<< HEAD
the updated Density-class object
=======
the updated Density-class object
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

See also

<<<<<<< HEAD

<<<<<<< HEAD ======= >>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1