S4 generic method to run analyses

This method carries out an analysis of distance sampling data. This method is provided to allow the user to perform diagnostics of the analyses used in the simulation. The data argument can be obtained by a call to simulate.survey(object, dht.table = TRUE). Note if the first object supplied is of class DS.Analysis then the second argument must be of class DDf.Data. The data argument may be of either class for an object argument of class Simulation.

Usage

analyse.data(analysis, data.obj, ...)

# S4 method for class 'DS.Analysis,Survey'
analyse.data(analysis, data.obj, warnings = NULL, ...)

# S4 method for class 'DS.Analysis,data.frame'
analyse.data(analysis, data.obj, warnings = NULL, transect = "line", ...)

Arguments

analysis

an object of class DS.Analysis

data.obj

an object of class Survey or a dataframe

...

optional arguments (currently not used)

warnings

a list of warnings and how many times they arose

transect

character value either "line" or "point" specifying type of transect used in survey

Value

a list containing an S3 ddf object and optionally an S3 dht object relating to the model with the minimum criteria.

either returns a list of the best model, warnings and the number of successfully fitted models (if warnings is supplied as a list) otherwise displays warnings as it goes and returns the best fitting ds model.