<<<<<<< HEAD

<<<<<<< HEAD
Distance Sampling Simulations 'dsims'
=======
Distance Sampling Simulations 'dsims'
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

Runs simulations of distance sampling surveys to help users optimise their survey designs for their particular study.

Details

The full process involves defining the study region, a description of the population of interest (including its distribution within the study region), a survey design, a detection process and one or more models to fit to the resulting data. The simulation engine will then use this information to generate both a population and a set of transects and simulate the detection process. The resulting data will be analysed and the estimates stored. By repeating this many times we can test the accuracy and precision of our estimates from various survey designs given our particular population of interest.

This package interfaces with the survey design package 'dssd' to create the survey regions, designs and generate the survey transects. While the 'DSsim' simulation package relied on survey transects already being contained in shapefiles within the supplied directory, dsims will generate the survey transects directly in R.

The main functions in this package are: make.density, make.population.description, make.detectability, make.ds.analysis, make.simulation, run.survey and run.simulation. See also make.region and make.design in the dssd package for examples of how to define study regions and designs.

Further information on distance sampling methods and example code is available at http://distancesampling.org/R/.

Also see our website for vignettes / example code at http://examples.distancesampling.org.

For help with distance sampling and this package, there is a Google Group https://groups.google.com/forum/#!forum/distance-sampling.

Author

Laura Marshall <lhm@st-and.ac.uk>

<<<<<<< HEAD

<<<<<<< HEAD ======= >>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1