<<<<<<< HEAD

<<<<<<< HEAD
plot
=======
plot
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

Produces four plots of the survey: 1) Plots the transects inside the survey region, 2) plots the population, 3) plots the transects, population and detections 4) plots a histogram of the detection distances. Note that only plots 3 & 4 are generated without the survey region if Region is omitted.

Usage
<<<<<<< HEAD

# S4 method for Survey,Region
plot(x, y, type = "all", ...)

# S4 method for Survey,ANY
=======
    

# S4 method for class 'Survey,Region'
plot(x, y, type = "all", ...)

# S4 method for class 'Survey,ANY'
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1
plot(x, y = NULL, type = "all", ...)

    

Arguments
<<<<<<< HEAD

x

object of class Survey

y

object of class Region or NULL

type
=======

x

object of class Survey

y

object of class Region or NULL

type
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

character specifies which plots you would like, defaults to "all". Other options include "transects", "population", "survey" and "distances". These will plot only the transects, only the population locations, both the transects and population with detections indicated or a histogram of the detection distances, respectively. Note that the final plots is only available if there were one or more detections. Only "survey" and "distances" available if the y Region argument is not supplied.

<<<<<<< HEAD
...
=======
...
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

additional plotting parameters

Value
<<<<<<< HEAD
Generate 4 plots showing the survey population, transects (including covered areas), detections and a histogram of the detection distances. Plots include the survey region. Also invisibly returns a list of ggplot objects if the user would like to customise the plots.

Generate 2 plots showing the survey population, transects (including covered areas), detections and a histogram of the detection distances. Plots do not include survey region. Also invisibly returns a list of ggplot objects if the user would like to customise the plots.

<<<<<<< HEAD ======= >>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1