Update cookies preferences

Plot

Plots an S4 object of class 'Detectability'

Usage

# S4 method for class 'Detectability,ANY'
plot(
  x,
  y,
  add = FALSE,
  plot.units = character(0),
  region.col = NULL,
  gap.col = NULL,
  main = "",
  ...
)

# S4 method for class 'Detectability,Population.Description'
plot(
  x,
  y,
  add = FALSE,
  plot.units = character(0),
  region.col = NULL,
  gap.col = NULL,
  main = "",
  ...
)

Arguments

x

object of class Detectability

y

object of class Population.Description

add

logical indicating whether it should be added to existing plot

plot.units

allows for units to be converted between m and km

region.col

fill colour for the region

gap.col

fill colour for the gaps

main

character plot title

...

other general plot parameters

Value

No return value, gives a warning to the user

No return value, plotting function