<<<<<<< HEAD

<<<<<<< HEAD
Method to run a simulation
=======
Method to run a simulation
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

Runs the simulation and returns the simulation object with results. If running in parallel and max.cores is not specified it will default to using one less than the number of cores / threads on your machine. For example code see make.simulation

Usage

run.simulation(
  simulation,
  run.parallel = FALSE,
  max.cores = NA,
  counter = TRUE,
  transect.path = character(0),
  progress.file = character(0)
)

Arguments
<<<<<<< HEAD

simulation

Simulation-class object

run.parallel

logical option to use multiple processors

max.cores
=======

simulation

Simulation-class object

run.parallel

logical option to use multiple processors

max.cores
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

integer maximum number of cores to use, if not specified then one less than the number available will be used.

<<<<<<< HEAD
counter

logical indicates if you would like to see the progress counter.

transect.path
=======
counter

logical indicates if you would like to see the progress counter.

transect.path
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

character gives the pathway to a folder of shapefiles or the path to a single shapefile (.shp file) which give the transects which should be used for the simulations. If a folder of transects a new shapefile will be used for each repetition. If a path specifying a single shapefile then the same transects will be used for each repetition.

<<<<<<< HEAD
progress.file
=======
progress.file
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

character path with filename to output progress to file for Distance for Windows progress counter. Not to be used when running directly in R.

Value
<<<<<<< HEAD
the Simulation-class object which now includes =======
the Simulation-class object which now includes >>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1 the results

See also

<<<<<<< HEAD

<<<<<<< HEAD ======= >>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1