S4 generic method to simulate a survey

Simulates the process by which individuals or clusters are detected. If a simulation is passed in then it will generate a population, set of transects and simulate the detection process. If a survey is passed in it will simply simulate the detection process. See make.simulation for example usage.

Usage

run.survey(object, ...)

# S4 method for class 'Simulation'
run.survey(object, filename = character(0))

# S4 method for class 'Survey.LT'
run.survey(object, region = NULL)

# S4 method for class 'Survey.PT'
run.survey(object, region = NULL)

Arguments

object

an object of class Simulation

...

allows extra arguments

filename

optional argument specifying a path to a shapefile if the transects are to be loaded from file.

region

an object of class Region.

Value

An object which inherits from a Survey-class object. This will be a Survey.LT-class object in the case of a simulation with a line transect design and a Survey.PT-class if the simulation has a point transect design.

See also