<<<<<<< HEAD

<<<<<<< HEAD
summary
=======
summary
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

Provides a summary of the simulation results.

Usage
<<<<<<< HEAD

# S4 method for Simulation
=======
    

# S4 method for class 'Simulation'
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1
summary(object, description.summary = TRUE, use.max.reps = FALSE, ...)

    

Arguments
<<<<<<< HEAD

object

object of class Simulation

description.summary
=======

object

object of class Simulation

description.summary
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

logical indicating whether an explanation of the summary should be displayed

<<<<<<< HEAD
use.max.reps

by default this is FALSE meaning that only simulation repetitions where all models converged for that data set are included. By setting this to TRUE any repetition where one or more models converged will be included in the summary results.

...
=======
use.max.reps

by default this is FALSE meaning that only simulation repetitions where all models converged for that data set are included. By setting this to TRUE any repetition where one or more models converged will be included in the summary results.

...
>>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1

no additional arguments currently implemented

Value
<<<<<<< HEAD
Object of class Simulation.Summary

<<<<<<< HEAD ======= >>>>>>> 27c21928aa011d2896717dce6fa50cb8563a8ac1