Update cookies preferences

Class "DS.Analysis"

Class "DDF.Analysis" is an S4 class describing a basic detection function model to be fitted to distance sampling data.

Slots

dfmodel

Object of class "formula"; describing the detection function model.

key

key function to use; "hn" gives half-normal (default), "hr" gives hazard-rate and "unif" gives uniform. Note that if uniform key is used, covariates cannot be included in the model.

adjustment

a list containing adjustment parameters: adjustment - either "cos" (recommended), "herm" or "poly", order - the orders of the adjustment terms to fit, scale - the scale by which the distances in the adjustment terms are divided. See details.

truncation

Object of class "list"; Specifies the truncation distance for the analyses.

cutpoints

Object of class "character"; gives the cutpoints of the bins for binned data analysis.

er.var

specifies which encounter rate variance estimator to use.

control.opts

A list to specify various options including monotonicity, method, initial.values.

group.strata

Dataframe with two columns ("design.id" and "analysis.id"). The former gives the strata names as defined in the design (i.e. the region object) the second specifies how they should be grouped (into less strata) for the analyses

criteria

Object of class "character"; describes which model selection criteria to use ("AIC","AICc","BIC").

Methods

run.analysis

signature=c(object = "DS.Analysis", data = data.frame): runs the analysis described in the object on the data provided.