Update cookies preferences

plot

Produces four plots of the survey: 1) Plots the transects inside the survey region, 2) plots the population, 3) plots the transects, population and detections 4) plots a histogram of the detection distances. Note that only plots 3 & 4 are generated without the survey region if Region is omitted.

Usage

# S4 method for class 'Survey,Region'
plot(x, y, type = "all", ...)

# S4 method for class 'Survey,ANY'
plot(x, y = NULL, type = "all", ...)

Arguments

x

object of class Survey

y

object of class Region or NULL

type

character specifies which plots you would like, defaults to "all". Other options include "transects", "population", "survey" and "distances". These will plot only the transects, only the population locations, both the transects and population with detections indicated or a histogram of the detection distances, respectively. Note that the final plots is only available if there were one or more detections. Only "survey" and "distances" available if the y Region argument is not supplied.

...

additional plotting parameters

Value

Generate 4 plots showing the survey population, transects (including covered areas), detections and a histogram of the detection distances. Plots include the survey region. Also invisibly returns a list of ggplot objects if the user would like to customise the plots.

Generate 2 plots showing the survey population, transects (including covered areas), detections and a histogram of the detection distances. Plots do not include survey region. Also invisibly returns a list of ggplot objects if the user would like to customise the plots.