Method to run a simulation

Runs the simulation and returns the simulation object with results. If running in parallel and max.cores is not specified it will default to using one less than the number of cores / threads on your machine. For example code see make.simulation

Usage

run.simulation(
  simulation,
  run.parallel = FALSE,
  max.cores = NA,
  counter = TRUE,
  transect.path = character(0),
  progress.file = character(0)
)

Arguments

simulation

Simulation-class object

run.parallel

logical option to use multiple processors

max.cores

integer maximum number of cores to use, if not specified then one less than the number available will be used.

counter

logical indicates if you would like to see the progress counter.

transect.path

character gives the pathway to a folder of shapefiles or the path to a single shapefile (.shp file) which give the transects which should be used for the simulations. If a folder of transects a new shapefile will be used for each repetition. If a path specifying a single shapefile then the same transects will be used for each repetition.

progress.file

character path with filename to output progress to file for Distance for Windows progress counter. Not to be used when running directly in R.

Value

the Simulation-class object which now includes the results

See also